requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
imutils==0.5.4
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
